+++ module/RallNic/view/RallNic/index/index.phtml
@@ -1,0 +1,1 @@
+<h1>Hello from RallNic's index view.</h1>
